feat: log natives with type information

+++ src/app/game/gta/Natives.hpp
@@ -3,6 +3,6 @@
 namespace gta::Natives
 {
     extern std::string_view GetNameByHash(uint64_t hash);
-    extern uint64_t GetHashByHandler(uint64_t handler);
-    extern std::unordered_map<uint64_t, uint64_t> GetAll();
+    extern uint64_t GetHashByHandler(uintptr_t handler);
+    extern std::unordered_map<uint64_t, uintptr_t> GetAll();
 }
